CHORE :: LoginAuth Type파일 생성
