+++ src/types/LoginAuth.type.ts
@@ -1,0 +1,9 @@
+export interface LoginAuth {
+  email: string;
+  password: string;
+}
+export interface ResetPassword {
+  code: string;
+  newPassword: string;
+  oldPassword: string;
+}
